SCHEDULING-856 Added display of nodes state
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@17573 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ rm-rcp/org.ow2.proactive.resourcemanager/src/org/ow2/proactive/resourcemanager/gui/topology/GraphLatencyFilter.java
@@ -151,7 +151,7 @@ public class GraphLatencyFilter extends GroupAction {
                 while (edgesIt.hasNext()) {
                     EdgeItem edge = (EdgeItem) (edgesIt.next());
                     NodeItem node = edge.getAdjacentItem(fixedNode);
-                    Double latency = (Double) (edge.get("weight"));
+                    Integer latency = (Integer) (edge.get("weight"));
                     if (!exclusive && (latency <= m_latency || node.isFixed())) {
                         visiblesNodes.add(node);
                     } else if (exclusive && latency > m_latency && !node.isFixed()) {
